refactor: Producer, Consumer -> interface, impl 형태로 변경 (#30)
